+++ include/memutils.h
@@ -8,6 +8,7 @@ extern "C" {
 
 void *memcpy(void *dst, const void *src, size_t n);
 void *memset(void *dst, int val, size_t n);
+int strncmp(const char *s1, const char *s2, size_t n);
 
 #ifdef __cplusplus
 }